# resources/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Order Management API",
    "description": "Server-to-server API for Laravel Order Management System with HMAC Authentication",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "{server_url}/api/v1",
      "description": "API Server",
      "variables": {
        "server_url": {
          "default": "http://localhost:8000",
          "description": "Base URL of the API server"
        }
      }
    }
  ],
  "security": [
    {
      "hmac_auth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "hmac_auth": {
        "type": "apiKey",
        "description": "HMAC-SHA256 authentication using custom headers. Required headers: X-Key-Id, X-Signature, X-Timestamp, Digest",
        "name": "X-Key-Id",
        "in": "header"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "Client": {
        "type": "object",
        "required": ["name", "email"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Client ID (ULID)",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "description": "Client name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Client email"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "description": "Client phone number"
          },
          "company": {
            "type": "string",
            "nullable": true,
            "description": "Company name"
          },
          "tax_id": {
            "type": "string",
            "nullable": true,
            "description": "Tax identification number"
          },
          "address": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Address"
              }
            ],
            "nullable": true,
            "description": "Client address"
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the client is active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
      "QueueStats": {
        "type": "object",
        "properties": {
          "pending": {
            "type": "integer",
            "description": "Number of pending jobs"
          },
          "failed": {
            "type": "integer",
            "description": "Number of failed jobs"
          },
          "recent": {
            "type": "integer",
            "description": "Number of recent jobs"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/orders": {
      "get": {
        "summary": "List Orders",
        "description": "Get a paginated list of orders",
        "tags": ["Orders"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 15
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by order status",
            "schema": {
              "type": "string",
              "enum": ["new", "confirmed", "paid", "fulfilled", "completed", "cancelled"]
            }
          },
          {
            "name": "client_id",
            "in": "query",
            "description": "Filter by client ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    },
                    "links": {
                      "type": "object",
                      "description": "Pagination links"
                    },
                    "meta": {
                      "type": "object",
                      "description": "Pagination metadata"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Order",
        "description": "Create a new order",
        "tags": ["Orders"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Order"
                  }
                ],
                "example": {
                  "number": "ORD-2024-001",
                  "client_id": "01HXYZ123456789ABCDEF",
                  "status": "new",
                  "total_amount": 99.99,
                  "currency": "EUR",
                  "shipping_address": {
                    "name": "John Doe",
                    "street": "123 Main St",
                    "city": "Prague",
                    "postal_code": "10000",
                    "country": "CZ",
                    "email": "john@example.com"
                  },
                  "billing_address": {
                    "name": "John Doe",
                    "street": "123 Main St",
                    "city": "Prague",
                    "postal_code": "10000",
                    "country": "CZ",
                    "email": "john@example.com"
                  },
                  "items": [
                    {
                      "name": "Sample Product",
                      "sku": "SKU-001",
                      "quantity": 2,
                      "unit_price": 49.99
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{orderId}": {
      "parameters": [
        {
          "name": "orderId",
          "in": "path",
          "required": true,
          "description": "Order ID (ULID)",
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "Get Order",
        "description": "Get a specific order by ID",
        "tags": ["Orders"],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Order",
        "description": "Update an existing order",
        "tags": ["Orders"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Order",
        "description": "Delete an order",
        "tags": ["Orders"],
        "responses": {
          "204": {
            "description": "Order deleted successfully"
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/orders/{orderId}/label": {
      "parameters": [
        {
          "name": "orderId",
          "in": "path",
          "required": true,
          "description": "Order ID (ULID)",
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "Generate Shipping Label",
        "description": "Generate a shipping label for an order",
        "tags": ["Orders", "Shipping"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "carrier": {
                    "type": "string",
                    "description": "Shipping carrier"
                  },
                  "service": {
                    "type": "string",
                    "description": "Shipping service"
                  },
                  "package_type": {
                    "type": "string",
                    "description": "Package type"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shipping label generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "label_id": {
                      "type": "string",
                      "description": "Label ID"
                    },
                    "tracking_number": {
                      "type": "string",
                      "description": "Tracking number"
                    },
                    "label_url": {
                      "type": "string",
                      "format": "uri",
                      "description": "URL to download label"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/orders/{orderId}/pdf": {
      "parameters": [
        {
          "name": "orderId",
          "in": "path",
          "required": true,
          "description": "Order ID (ULID)",
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "Generate PDF",
        "description": "Generate a PDF document for an order (rate limited)",
        "tags": ["Orders", "PDF"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "template": {
                    "type": "string",
                    "description": "PDF template to use"
                  },
                  "format": {
                    "type": "string",
                    "enum": ["A4", "Letter"],
                    "description": "PDF page format"
                  },
                  "orientation": {
                    "type": "string",
                    "enum": ["portrait", "landscape"],
                    "description": "PDF orientation"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "PDF generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pdf_url": {
                      "type": "string",
                      "format": "uri",
                      "description": "URL to download PDF"
                    },
                    "pdf_path": {
                      "type": "string",
                      "description": "PDF file path"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/clients": {
      "get": {
        "summary": "List Clients",
        "description": "Get a paginated list of clients",
        "tags": ["Clients"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 15
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of clients",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Client"
                      }
                    },
                    "links": {
                      "type": "object",
                      "description": "Pagination links"
                    },
                    "meta": {
                      "type": "object",
                      "description": "Pagination metadata"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Client",
        "description": "Create a new client",
        "tags": ["Clients"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              },
              "example": {
                "name": "Acme Corporation",
                "email": "contact@acme.com",
                "phone": "+420123456789",
                "company": "Acme Corporation",
                "tax_id": "CZ12345678",
                "address": {
                  "name": "Acme Corporation",
                  "street": "Business District 123",
                  "city": "Prague",
                  "postal_code": "11000",
                  "country": "CZ",
                  "email": "contact@acme.com"
                },
                "is_active": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Client created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/queues/stats": {
      "get": {
        "summary": "Queue Statistics",
        "description": "Get queue statistics including pending and failed job counts",
        "tags": ["Queue"],
        "responses": {
          "200": {
            "description": "Queue statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueStats"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}